Fix negative coordinate mappings in `get_block`
Change local block coordinate parameters from int to Uint8 across WorldMap API and implementations. Revise global->chunk and local coordinate calculations to correctly handle negative global coordinates (fix division/modulo logic and offsets). Add debug logging and test chunk creation in GameWorld::generate_world_map to exercise negative chunk indices and various local/global lookups. Note: chunk-based get_block/set_block bodies remain TODO; added temporary logging/tests to validate mapping behavior.

+++ include/WorldMap.hpp
@@ -147,14 +147,14 @@ public:
 
   // Block retrieval functions
   BlockID get_block(const ChunkPos& chunk_pos,
-                    int local_x, int local_y, int local_z) const;
+                    Uint8 local_x, Uint8 local_y, Uint8 local_z) const;
   
   BlockID get_block(Sint64 global_x, Sint64 global_y, Sint64 global_z) const;
 
 
   // Block setting functions
   bool set_block(const ChunkPos& chunk_pos,
-                 int local_x, int local_y, int local_z,
+                 Uint8 local_x, Uint8 local_y, Uint8 local_z,
                  BlockID block_id);
 
   bool set_block(Sint64 global_x, Sint64 global_y, Sint64 global_z,
